Deconstruction of d3drender for skybox
