+++ clientd3d/d3drender.c
@@ -54,8 +54,6 @@ LPDIRECT3DTEXTURE9		gpBackBufferTexFull;
 LPDIRECT3DTEXTURE9		gpViewElements[NUM_VIEW_ELEMENTS];
 LPDIRECT3DTEXTURE9		gpSunTex;
 
-LPDIRECT3DTEXTURE9		gpSkyboxTextures[5][6];
-
 D3DVIEWPORT9			gViewport;
 D3DCAPS9				gD3DCaps;
 
@@ -102,7 +100,6 @@ PALETTEENTRY			gPalette[256];
 unsigned int			gFrame = 0;
 int						gScreenWidth;
 int						gScreenHeight;
-int						gCurBackground;
 
 // The size of the main full size render buffer and also a smaller buffer for effects.
 // The smaller buffer is used for effects that don't need full resolution.
@@ -165,119 +162,11 @@ extern BYTE				light_rows[MAXY/2+1];      // Strength of light as function of sc
 extern ViewElement		ViewElements[];
 extern HDC				gBitsDC;
 
-#define SKYBOX_DIMENSIONS	(75000.0f)
-
-float gSkyboxXYZ[] =
-{
-	// back
-	SKYBOX_DIMENSIONS, 37000.0f, -SKYBOX_DIMENSIONS,
-	SKYBOX_DIMENSIONS, -37000.0f, -SKYBOX_DIMENSIONS,
-	-SKYBOX_DIMENSIONS, -37000.0f, -SKYBOX_DIMENSIONS,
-	-SKYBOX_DIMENSIONS, 37000.0f, -SKYBOX_DIMENSIONS,
-
-	// bottom
-	-SKYBOX_DIMENSIONS, -37000.0f, SKYBOX_DIMENSIONS,
-	-SKYBOX_DIMENSIONS, -37000.0f, -SKYBOX_DIMENSIONS,
-	SKYBOX_DIMENSIONS, -37000.0f, -SKYBOX_DIMENSIONS,
-	SKYBOX_DIMENSIONS, -37000.0f, SKYBOX_DIMENSIONS,
-
-	// front
-	-SKYBOX_DIMENSIONS, 37000.0f, SKYBOX_DIMENSIONS,
-	-SKYBOX_DIMENSIONS, -37000.0f, SKYBOX_DIMENSIONS,
-	SKYBOX_DIMENSIONS, -37000.0f, SKYBOX_DIMENSIONS,
-	SKYBOX_DIMENSIONS, 37000.0f, SKYBOX_DIMENSIONS,
-
-	// left
-	-SKYBOX_DIMENSIONS, 37000.0f, -SKYBOX_DIMENSIONS,
-	-SKYBOX_DIMENSIONS, -37000.0f, -SKYBOX_DIMENSIONS,
-	-SKYBOX_DIMENSIONS, -37000.0f, SKYBOX_DIMENSIONS,
-	-SKYBOX_DIMENSIONS, 37000.0f, SKYBOX_DIMENSIONS,
-
-	// right
-	SKYBOX_DIMENSIONS, 37000.0f, SKYBOX_DIMENSIONS,
-	SKYBOX_DIMENSIONS, -37000.0f, SKYBOX_DIMENSIONS,
-	SKYBOX_DIMENSIONS, -37000.0f, -SKYBOX_DIMENSIONS,
-	SKYBOX_DIMENSIONS, 37000.0f, -SKYBOX_DIMENSIONS,
-
-	// top
-	-SKYBOX_DIMENSIONS, 37000.0f, -SKYBOX_DIMENSIONS,
-	-SKYBOX_DIMENSIONS, 37000.0f, SKYBOX_DIMENSIONS,
-	SKYBOX_DIMENSIONS, 37000.0f, SKYBOX_DIMENSIONS,
-	SKYBOX_DIMENSIONS, 37000.0f, -SKYBOX_DIMENSIONS,
-};
-
-float gSkyboxST[] =
-{
-	0.001f, 0.001f,
-	0.001f, 0.999f,
-	0.999f, 0.999f,
-	0.999f, 0.001f,
-
-	0.001f, 0.001f,
-	0.001f, 0.999f,
-	0.999f, 0.999f,
-	0.999f, 0.001f,
-
-	0.001f, 0.001f,
-	0.001f, 0.999f,
-	0.999f, 0.999f,
-	0.999f, 0.001f,
-
-	0.001f, 0.001f,
-	0.001f, 0.999f,
-	0.999f, 0.999f,
-	0.999f, 0.001f,
-
-	0.001f, 0.001f,
-	0.001f, 0.999f,
-	0.999f, 0.999f,
-	0.999f, 0.001f,
-
-	0.001f, 0.001f,
-	0.001f, 0.999f,
-	0.999f, 0.999f,
-	0.999f, 0.001f,
-};
-
-unsigned char gSkyboxBGRA[] =
-{
-	192, 192, 192, 255,
-	192, 192, 192, 255,
-	192, 192, 192, 255,
-	192, 192, 192, 255,
-
-	192, 192, 192, 255,
-	192, 192, 192, 255,
-	192, 192, 192, 255,
-	192, 192, 192, 255,
-
-	192, 192, 192, 255,
-	192, 192, 192, 255,
-	192, 192, 192, 255,
-	192, 192, 192, 255,
-
-	192, 192, 192, 255,
-	192, 192, 192, 255,
-	192, 192, 192, 255,
-	192, 192, 192, 255,
-
-	192, 192, 192, 255,
-	192, 192, 192, 255,
-	192, 192, 192, 255,
-	192, 192, 192, 255,
-
-	192, 192, 192, 255,
-	192, 192, 192, 255,
-	192, 192, 192, 255,
-	192, 192, 192, 255,
-};
-
-void				D3DRenderBackgroundsLoad(char *pFilename, int index);
 
 void				D3DRenderLMapsBuild(void);
 void				D3DLMapsStaticGet(room_type *room);
 void				D3DRenderFontInit(font_3d *pFont, HFONT hFont);
-void				D3DRenderSkyboxDraw(d3d_render_pool_new *pPool, int angleHeading, int anglePitch);
+
 void				D3DRenderBackgroundOverlays(d3d_render_pool_new* pPool, int angleHeading, int anglePitch, room_type* room, Draw3DParams* params);
 void				D3DRenderSunDraw(int angleHeading, int anglePitch);
 
@@ -475,7 +364,7 @@ HRESULT D3DRenderInit(HWND hWnd)
 
 void D3DRenderShutDown(void)
 {
-	int	i, j;
+	int	i;
 
 	if (!gD3DDriverProfile.bSoftwareRenderer)
 	{
@@ -527,15 +416,7 @@ void D3DRenderShutDown(void)
 			gpBackBufferTex[i] = NULL;
 		}
 
-		for (j = 0; j < 5; j++)
-			for (i = 0; i < 6; i++)
-			{
-				if (gpSkyboxTextures[j][i])
-				{
-               IDirect3DTexture9_Release(gpSkyboxTextures[j][i]);
-					gpSkyboxTextures[j][i] = NULL;
-				}
-			}
+		D3DRenderSkyBoxShutdown();
 
 		for (i = 0; i < NUM_VIEW_ELEMENTS; i++)
 		{
@@ -571,7 +452,6 @@ void D3DRenderBegin(room_type *room, Draw3DParams *params)
 	int			angleHeading, anglePitch;
 	int			curPacket = 0;
 	int			curIndex = 0;
-	static ID	tempBkgnd = 0;
 	room_contents_node *pRNode = nullptr;
 
 	timeOverall = timeGetTime();
@@ -584,19 +464,8 @@ void D3DRenderBegin(room_type *room, Draw3DParams *params)
 	Bool draw_objects = can_see;
 	Bool draw_particles = can_see;
 	Bool draw_background_overlays = can_see;
-	if (gpSkyboxTextures[0][0] == NULL)
-	{
-		D3DRenderBackgroundsLoad("./resource/skya.bsf", 0);
-		D3DRenderBackgroundsLoad("./resource/skyb.bsf", 1);
-		D3DRenderBackgroundsLoad("./resource/skyc.bsf", 2);
-		D3DRenderBackgroundsLoad("./resource/skyd.bsf", 3);
-		D3DRenderBackgroundsLoad("./resource/redsky.bsf", 4);
-	}
-	if (tempBkgnd != current_room.bkgnd)
-	{
-		D3DRenderBackgroundSet(current_room.bkgnd);
-		tempBkgnd = current_room.bkgnd;
-	}
+
+	D3DRenderSkyBoxBegin();
 
 	// view element textures
 	if (gFrame == 0)
@@ -675,46 +544,7 @@ void D3DRenderBegin(room_type *room, Draw3DParams *params)
 
 	if (draw_sky) // Render the skybox first
 	{
-		// Set render states for skybox
-		IDirect3DDevice9_SetRenderState(gpD3DDevice, D3DRS_CULLMODE, D3DCULL_NONE);
-		SetZBias(gpD3DDevice, 0);
-		IDirect3DDevice9_SetRenderState(gpD3DDevice, D3DRS_ZWRITEENABLE, FALSE);
-		IDirect3DDevice9_SetRenderState(gpD3DDevice, D3DRS_ZENABLE, FALSE);
-
-		// Disable alpha blending and alpha testing for the skybox
-		IDirect3DDevice9_SetRenderState(gpD3DDevice, D3DRS_ALPHABLENDENABLE, FALSE);
-		IDirect3DDevice9_SetRenderState(gpD3DDevice, D3DRS_ALPHATESTENABLE, FALSE);
-
-		// Disable fog for the skybox
-		IDirect3DDevice9_SetRenderState(gpD3DDevice, D3DRS_FOGENABLE, FALSE);
-
-		// Set texture stages for the skybox
-		D3DRENDER_SET_COLOR_STAGE(gpD3DDevice, 0, D3DTOP_SELECTARG1, D3DTA_TEXTURE, D3DTA_DIFFUSE);
-		D3DRENDER_SET_ALPHA_STAGE(gpD3DDevice, 0, D3DTOP_SELECTARG1, D3DTA_TEXTURE, D3DTA_DIFFUSE);
-		D3DRENDER_SET_COLOR_STAGE(gpD3DDevice, 1, D3DTOP_DISABLE, 0, 0);
-		D3DRENDER_SET_ALPHA_STAGE(gpD3DDevice, 1, D3DTOP_DISABLE, 0, 0);
-
-		// Set vertex shader and declaration for the skybox
-		IDirect3DDevice9_SetVertexShader(gpD3DDevice, NULL);
-		IDirect3DDevice9_SetVertexDeclaration(gpD3DDevice, decl1dc);
-
-		// Render the skybox
-		D3DRenderPoolReset(&gWorldPool, &D3DMaterialWorldPool);
-		D3DRenderSkyboxDraw(&gWorldPool, angleHeading, anglePitch);
-		D3DCacheFill(&gWorldCacheSystem, &gWorldPool, 1);
-		D3DCacheFlush(&gWorldCacheSystem, &gWorldPool, 1, D3DPT_TRIANGLESTRIP);
-
-		// Restore render states after skybox rendering
-		IDirect3DDevice9_SetRenderState(gpD3DDevice, D3DRS_ZWRITEENABLE, TRUE);
-		IDirect3DDevice9_SetRenderState(gpD3DDevice, D3DRS_ZENABLE, TRUE);
-		IDirect3DDevice9_SetRenderState(gpD3DDevice, D3DRS_FOGENABLE, TRUE);
-
-		// Restore alpha blending and alpha testing for subsequent rendering
-		IDirect3DDevice9_SetRenderState(gpD3DDevice, D3DRS_ALPHABLENDENABLE, TRUE);
-		IDirect3DDevice9_SetRenderState(gpD3DDevice, D3DRS_ALPHATESTENABLE, TRUE);
-
-		// restore the correct view matrix
-		IDirect3DDevice9_SetTransform(gpD3DDevice, D3DTS_VIEW, &view);
+		D3DRenderSkyBox(room, params, pRNode, angleHeading, anglePitch);
 	}
 
 	if (gD3DRedrawAll & D3DRENDER_REDRAW_ALL)
@@ -1809,74 +1639,6 @@ void D3DRenderBackgroundOverlays(d3d_render_pool_new* pPool, int angleHeading, i
 	}
 }
 
-void D3DRenderSkyboxDraw(d3d_render_pool_new *pPool, int angleHeading, int anglePitch)
-{
-	int			i, j;
-	D3DMATRIX	rot, mat;
-
-	d3d_render_packet_new	*pPacket;
-	d3d_render_chunk_new	*pChunk;
-
-	IDirect3DDevice9_SetRenderState(gpD3DDevice, D3DRS_ALPHABLENDENABLE, FALSE);
-
-	MatrixIdentity(&mat);
-	IDirect3DDevice9_SetTransform(gpD3DDevice, D3DTS_WORLD, &mat);
-
-	MatrixRotateY(&rot, (float)angleHeading * 360.0f / 4096.0f * PI / 180.0f);
-	MatrixRotateX(&mat, (float)anglePitch * 45.0f / 414.0f * PI / 180.0f);
-	MatrixMultiply(&mat, &rot, &mat);
-
-	IDirect3DDevice9_SetTransform(gpD3DDevice, D3DTS_VIEW, &mat);
-
-	IDirect3DDevice9_SetSamplerState(gpD3DDevice, 0,
-                D3DSAMP_ADDRESSU, D3DTADDRESS_CLAMP);
-	IDirect3DDevice9_SetSamplerState(gpD3DDevice, 0,
-				D3DSAMP_ADDRESSV, D3DTADDRESS_CLAMP);
-
-	for (i = 0; i < 6; i++)
-	{
-		pPacket = D3DRenderPacketFindMatch(pPool, gpSkyboxTextures[gCurBackground][i], NULL, 0, 0, 0);
-		if (NULL == pPacket)
-			return;
-
-		pChunk = D3DRenderChunkNew(pPacket);
-		assert(pChunk);
-
-		pChunk->numIndices = 4;
-		pChunk->numVertices = 4;
-		pChunk->numPrimitives = pChunk->numVertices - 2;
-		pPacket->pMaterialFctn = &D3DMaterialWorldPacket;
-		pChunk->pMaterialFctn = &D3DMaterialWorldDynamicChunk;
-		pChunk->flags |= D3DRENDER_NOAMBIENT;
-
-		// add xyz, st, and bgra data
-		for (j = 0; j < 4; j++)
-		{
-			pChunk->xyz[j].x = gSkyboxXYZ[(i * 4 * 3) + (j * 3)];
-			pChunk->xyz[j].z = gSkyboxXYZ[(i * 4 * 3) + (j * 3) + 1];
-			pChunk->xyz[j].y = gSkyboxXYZ[(i * 4 * 3) + (j * 3) + 2];
-
-			pChunk->st0[j].s = gSkyboxST[(i * 4 * 2) + (j * 2)];
-			pChunk->st0[j].t = gSkyboxST[(i * 4 * 2) + (j * 2) + 1];
-
-			pChunk->bgra[j].b = gSkyboxBGRA[(i * 4 * 4) + (j * 4)];
-			pChunk->bgra[j].g = gSkyboxBGRA[(i * 4 * 4) + (j * 4) + 1];
-			pChunk->bgra[j].r = gSkyboxBGRA[(i * 4 * 4) + (j * 4) + 2];
-			pChunk->bgra[j].a = gSkyboxBGRA[(i * 4 * 4) + (j * 4) + 3];
-
-		}
-		pChunk->indices[0] = 1;
-		pChunk->indices[1] = 2;
-		pChunk->indices[2] = 0;
-		pChunk->indices[3] = 3;
-	}
-
-	IDirect3DDevice9_SetSamplerState(gpD3DDevice, 0,
-				D3DSAMP_ADDRESSU, D3DTADDRESS_WRAP);
-	IDirect3DDevice9_SetSamplerState(gpD3DDevice, 0,
-				D3DSAMP_ADDRESSV, D3DTADDRESS_WRAP);
-}
-
 int D3DRenderWallExtract(WallData *pWall, PDIB pDib, unsigned int *flags, custom_xyz *pXYZ,
 						  custom_st *pST, custom_bgra *pBGRA, unsigned int type, int side)
 {
@@ -2819,138 +2581,6 @@ void D3DRenderCeilingExtract(BSPnode *pNode, PDIB pDib, custom_xyz *pXYZ, custom
 	}
 }
 
-void D3DRenderBackgroundsLoad(char *pFilename, int index)
-{
-	FILE	*pFile;
-	png_structp	pPng = NULL;
-	png_infop	pInfo = NULL;
-	png_infop	pInfoEnd = NULL;
-	png_bytepp   rows;
-
-	D3DLOCKED_RECT		lockedRect;
-	unsigned char		*pBits = NULL;
-	unsigned int		w, h, b;
-	int					pitchHalf, bytePP;
-	fpos_t				pos;
-
-	pFile = fopen(pFilename, "rb");
-	if (pFile == NULL)
-		return;
-
-	pPng = png_create_read_struct(PNG_LIBPNG_VER_STRING, NULL, NULL, NULL);
-	if (NULL == pPng)
-	{
-		fclose(pFile);
-		return;
-	}
-
-	pInfo = png_create_info_struct(pPng);
-	if (NULL == pInfo)
-	{
-		png_destroy_read_struct(&pPng, NULL, NULL);
-		fclose(pFile);
-		return;
-	}
-
-	pInfoEnd = png_create_info_struct(pPng);
-	if (NULL == pInfoEnd)
-	{
-		png_destroy_read_struct(&pPng, &pInfo, NULL);
-		fclose(pFile);
-		return;
-	}
-
-	if (setjmp(png_jmpbuf(pPng)))
-	{
-		png_destroy_read_struct(&pPng, &pInfo, &pInfoEnd);
-		fclose(pFile);
-		return;
-	}
-
-	png_destroy_read_struct(&pPng, &pInfo, &pInfoEnd);
-
-	pos = 0;
-
-	{
-		int	i;
-		png_bytep	curRow;
-
-		for (i = 0; i < 6; i++)
-		{
-			pPng = png_create_read_struct(PNG_LIBPNG_VER_STRING, NULL, NULL, NULL);
-			pInfo = png_create_info_struct(pPng);
-			pInfoEnd = png_create_info_struct(pPng);
-			setjmp(png_jmpbuf(pPng));
-
-			fseek(pFile, pos, SEEK_SET);
-
-			png_init_io(pPng, pFile);
-			png_read_png(pPng, pInfo, PNG_TRANSFORM_IDENTITY, NULL);
-			rows = png_get_rows(pPng, pInfo);
-
-      unsigned int image_width = png_get_image_width(pPng, pInfo);
-      unsigned int image_height = png_get_image_height(pPng, pInfo);
-      bytePP = png_get_bit_depth(pPng, pInfo) / 8;
-
-			IDirect3DDevice9_CreateTexture(gpD3DDevice, image_width, image_height, 1, 0,
-                                     D3DFMT_A8R8G8B8, D3DPOOL_MANAGED, &gpSkyboxTextures[index][i], NULL);
-
-			IDirect3DTexture9_LockRect(gpSkyboxTextures[index][i], 0, &lockedRect, NULL, 0);
-
-			pitchHalf = lockedRect.Pitch / 2;
-
-			pBits = (unsigned char *)lockedRect.pBits;
-
-			for (h = 0; h < image_height; h++)
-			{
-				curRow = rows[h];
-
-				for (w = 0; w < image_width; w++)
-				{
-					for (b = 0; b < 4; b++)
-					{
-						if (b == 3)
-							pBits[h * lockedRect.Pitch + w * 4 + b] = 255;
-						else
-              // Apparently PNGs are BGR, while DirectX wants RGB
-							pBits[h * lockedRect.Pitch + w * 4 + (2 - b)] =
-								curRow[(w * 3) + b];
-					}
-				}
-			}
-
-			IDirect3DTexture9_UnlockRect(gpSkyboxTextures[index][i], 0);
-
-			fgetpos(pFile, &pos);
-			png_destroy_read_struct(&pPng, &pInfo, &pInfoEnd);
-		}
-	}
-
-	fclose(pFile);
-}
-
-void D3DRenderBackgroundSet(ID background)
-{
-	char	*filename = NULL;
-	filename = LookupRsc(background);
-
-	if (filename)
-	{
-		if (strstr(filename, "skya.bgf"))
-			gCurBackground = 0;
-		else if (strstr(filename, "skyb.bgf"))
-			gCurBackground = 1;
-		else if (strstr(filename, "skyc.bgf"))
-			gCurBackground = 2;
-		else if (strstr(filename, "skyd.bgf"))
-			gCurBackground = 3;
-		else if (strstr(filename, "redsky.bgf"))
-			gCurBackground = 4;
-	}
-
-	// force a rebuild since static lightmaps might have changed
-	gD3DRedrawAll |= D3DRENDER_REDRAW_ALL;
-}
 
 // new render stuff
 void D3DRenderPoolInit(d3d_render_pool_new *pPool, int size, int packetSize)
